+++ lib/src/route.dart
@@ -1,8 +1,8 @@
 import 'package:dartness/src/callable.dart';
 import 'package:dartness/src/middleware.dart';
 import 'package:dartness/src/context.dart';
+import 'package:dartness/src/routeMatcher.dart';
 
-import 'dart:math';
 import 'dart:async';
 
 class Route {
@@ -11,39 +11,19 @@ class Route {
   Function callback;
   final List<Callable> _before = [];
   final List<Callable> _after = [];
-
+  bool useRegexp = false;
   Map<String, String> params = {};
 
-  Route(this.method, this.path, this.callback);
+  Route(this.method, this.path, this.callback, {this.useRegexp = false});
 
   bool isMatching(String method, Uri uri) {
     if (this.method != method && this.method != '*') {
       return false;
     }
-    final pathParts = path.split('/');
-    final uriPaths = uri.path.split('/');
-
-    if (pathParts.length != uriPaths.length) {
-      return false;
-    }
-
-    final len = min(pathParts.length, uriPaths.length);
-
-    for (var q = 0; q < len; q++) {
-      if (pathParts[q].startsWith(':')) {
-        final key = pathParts[q].substring(1);
-        params[key] = uriPaths[q];
-        continue;
-      }
-
-      if (pathParts[q] == uriPaths[q]) {
-        continue;
-      } else {
-        return false;
-      }
-    }
-
-    return true;
+    final matcher = new RouteMatcher();
+    matcher.match(path, uri.path);
+    params = matcher.params;
+    return matcher.matched;
   }
 
   Future<void> call(Context context) async {