+++ CHANGELOG.md
@@ -5,6 +5,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+## [0.1.0] - 2018-09-09
+
+### Added
+- added regexp routes
+
+### Changed
+- corrected deprecated
+- corrected docs
 
 ## [0.0.18] - 2018-09-07
 ### Added